polyline generator refactor
