+++ shaders/drawable.line_basic.fragment.glsl
@@ -17,7 +17,13 @@ in float v_width;
 in vec2 v_normal;
 in float v_gamma_scale;
 
+#pragma mapbox: define highp vec4 color
+#pragma mapbox: define lowp float opacity
+
 void main() {
+    #pragma mapbox: initialize highp vec4 color
+    #pragma mapbox: initialize lowp float opacity
+
     // Calculate the distance of the pixel from the line in pixels.
     float dist = length(v_normal) * v_width;
 
@@ -27,7 +33,7 @@ void main() {
     float blur2 = (1.0 / u_device_pixel_ratio) * v_gamma_scale;
     float alpha = clamp(min(dist + blur2, v_width - dist) / blur2, 0.0, 1.0);
 
-    fragColor = u_color * (alpha * u_opacity);
+    fragColor = color * (alpha * opacity);
 
 #ifdef OVERDRAW_INSPECTOR
     fragColor = vec4(1.0);
