+++ src/mbgl/gfx/fill_generator.cpp
@@ -154,17 +154,7 @@ void generateFillAndOutineBuffers(const GeometryCollection& geometry,
                                   gfx::VertexVector<LineLayoutVertex>& lineVertices,
                                   gfx::IndexVector<gfx::Triangles>& lineIndexes,
                                   SegmentVector<LineAttributes>& lineSegments) {
-    gfx::PolylineGenerator<LineLayoutVertex, Segment<LineAttributes>> lineGenerator(
-        lineVertices,
-        LineProgram::layoutVertex,
-        lineSegments,
-        [](std::size_t vertexOffset, std::size_t indexOffset) -> Segment<LineAttributes> {
-            return Segment<LineAttributes>(vertexOffset, indexOffset);
-        },
-        [](auto& seg) -> Segment<LineAttributes>& { return seg; },
-        lineIndexes);
-
-    gfx::PolylineGeneratorOptions lineOptions;
+    gfx::PolylineGenerator::Options lineOptions;
     lineOptions.type = FeatureType::Polygon;
 
     for (auto& polygon : classifyRings(geometry)) {
@@ -176,7 +166,17 @@ void generateFillAndOutineBuffers(const GeometryCollection& geometry,
 
         for (const auto& ring : polygon) {
             addRingVertices(fillVertices, ring);
-            lineGenerator.generate(ring, lineOptions);
+            gfx::PolylineGenerator::generate(
+                lineVertices,
+                LineProgram::layoutVertex,
+                lineSegments,
+                [](std::size_t vertexOffset, std::size_t indexOffset) -> Segment<LineAttributes> {
+                    return Segment<LineAttributes>(vertexOffset, indexOffset);
+                },
+                [](auto& seg) -> Segment<LineAttributes>& { return seg; },
+                lineIndexes,
+                ring,
+                lineOptions);
         }
 
         std::vector<uint32_t> indices = mapbox::earcut(polygon);
@@ -193,17 +193,7 @@ void generateFillAndOutineBuffers(const GeometryCollection& geometry,
                                   SegmentVector<LineAttributes>& lineSegments,
                                   gfx::IndexVector<gfx::Lines>& basicLineIndexes,
                                   SegmentVector<FillAttributes>& basicLineSegments) {
-    gfx::PolylineGenerator<LineLayoutVertex, Segment<LineAttributes>> lineGenerator(
-        lineVertices,
-        LineProgram::layoutVertex,
-        lineSegments,
-        [](std::size_t vertexOffset, std::size_t indexOffset) -> Segment<LineAttributes> {
-            return Segment<LineAttributes>(vertexOffset, indexOffset);
-        },
-        [](auto& seg) -> Segment<LineAttributes>& { return seg; },
-        lineIndexes);
-
-    gfx::PolylineGeneratorOptions lineOptions;
+    gfx::PolylineGenerator::Options lineOptions;
     lineOptions.type = FeatureType::Polygon;
 
     for (auto& polygon : classifyRings(geometry)) {
@@ -217,7 +207,17 @@ void generateFillAndOutineBuffers(const GeometryCollection& geometry,
             std::size_t base = fillVertices.elements();
             std::size_t nVertices = addRingVertices(fillVertices, ring);
             addOutlineIndices(base, nVertices, basicLineSegments, basicLineIndexes);
-            lineGenerator.generate(ring, lineOptions);
+            gfx::PolylineGenerator::generate(
+                lineVertices,
+                LineProgram::layoutVertex,
+                lineSegments,
+                [](std::size_t vertexOffset, std::size_t indexOffset) -> Segment<LineAttributes> {
+                    return Segment<LineAttributes>(vertexOffset, indexOffset);
+                },
+                [](auto& seg) -> Segment<LineAttributes>& { return seg; },
+                lineIndexes,
+                ring,
+                lineOptions);
         }
 
         std::vector<uint32_t> indices = mapbox::earcut(polygon);
